mamba mvebu: enable overlay on syscfg partition formatted as ubifs
Mamba use the syscfg partition formatted in ubifs type as the overlay
partition to store the router configuration. This overlay style is
different from OpenWRT. To avoid affecting the whole OpenWRT tree, we do
not modify OpenWRT's block-mount package. Instead, we add a script to
carry out the overlay on syscfg partition

Signed-off-by: Matthew Fatheree <Matthew.Fatheree@belkin.com>

Conflicts:

	target/linux/mvebu/config-3.10

Signed-off-by: Matthew Fatheree <Matthew.Fatheree@belkin.com>