mamba mvebu: sysupgrade support for mamba + dual boot is supportted
Add platform specific support for sysupgrade. Mamba supports dual boot
implementation. The firmware can be flashed to Primary Boot partition or
Alternative Boot partition. The sysupgrade has the ping-pong effect:
when we are runnning the firmware from Primary partition and do
sysupgrade the new firmware is written to the Alternative partition, on
next reboot bootcmd is set to Alternative partition; when we are running
the firmware from Alternative partition and do sysupgrade, the new
firmware is written to the Primary partition, on next reboot bootcmd is
set to Primary partition. boot_part is used to distinguished between
Primary Boot partition and Alternative Boot partition

Signed-off-by: Matthew Fatheree <Matthew.Fatheree@belkin.com>