mamba mvebu: support factory reset on syscfg partition
Mamba use the syscfg partition formatted in ubifs type as the
overlay partition to store the router configuration. This overlay style
is different from OpenWRT. We created a jffs2reset_mamba script to
support resetting this overlay partition. The original jffs2reset on the
rootfs is  a symbolic link to /sbin/mount_root, we cannot overwrite the
symbolic link directly. We created the script with a different name
jffs2reset_mamba. During the Image/mkfs/prepare/default, we try to
overwrite the jffs2reset with jffs2reset_mamba

system.lua_mamba file is added to support the factory reset from UI, it only
changes the default IP address to 192.168.200.1 and checking condition
on the overlay partition to enable the "Perform reset" button.
system.lua is installed by luci, we overwrite system.lua with
system.lua_mamba

Signed-off-by: Matthew Fatheree <Matthew.Fatheree@belkin.com>