mamba mvebu: support switch2jffs for mamba
Mamba use the syscfg partition formatted in ubifs type as the overlay
partition to store the router configuration. This overlay style is
different from OpenWRT. We created a swith2jffs_mamba script to support
mounting the overlayfs in Fail Safe Mode or in case of ramoverlay. The
original switch2jffs on the rootfs is  a symbolic link to
/sbin/mount_root, we cannot overwrite the symbolic link directly. We
created the script with a different name - switch2jffs_mamba. During the
Image/mkfs/prepare/default, we try to overwrite the switch2jffs with
switch2jffs_mamba

Signed-off-by: Matthew Fatheree <Matthew.Fatheree@belkin.com>